// packages/opencode/src/file/ripgrep.d.ts
export declare namespace Ripgrep {
    const Begin: z.ZodObject<{
        type: z.ZodLiteral<"begin">;
        data: z.ZodObject<{
            path: z.ZodObject<{
                text: z.ZodString;
            }, z.core.$strip>;
        }, z.core.$strip>;
    }, z.core.$strip>;
// ...
}